added docs, cleaned up
