+++ exp/txnbuild/inflation.go
@@ -1,35 +1,22 @@
 package txnbuild
 
 import (
-	"github.com/stellar/go/support/errors"
 	"github.com/stellar/go/xdr"
 )
 
-// Inflation ...
+// Inflation represents the Stellar inflation operation. See
+// https://www.stellar.org/developers/guides/concepts/list-of-operations.html
 type Inflation struct {
 	xdrOp struct{}
 }
 
-// Build ...
-func (inf *Inflation) Build() (xdr.Operation, error) {
-	// Create operation body
-	body, err := inf.NewXDROperationBody()
-	if err != nil {
-		return xdr.Operation{}, errors.Wrap(err, "Failed to create XDR")
-	}
-	// Append relevant operation to TX.operations
-	xdrOperation := xdr.Operation{Body: body}
-
-	return xdrOperation, nil
+// Init for Inflation is a no-op to match the txnbuild.Operation interface.
+func (inf *Inflation) Init() error {
+	return nil
 }
 
-// NewXDROperationBody ...
+// NewXDROperationBody for Inflation initialises the corresponding XDR body.
 func (inf *Inflation) NewXDROperationBody() (xdr.OperationBody, error) {
-	// TODO: Better name
-	// TODO: Add next two lines in here
-
 	opType := xdr.OperationTypeInflation
-	body, err := xdr.NewOperationBody(opType, nil)
-
-	return body, err
+	return xdr.NewOperationBody(opType, nil)
 }
